refactor: comment CRUD auth 없이 user 연결

+++ src/main/java/kpl/fiml/comment/application/CommentService.java
@@ -5,6 +5,8 @@ import kpl.fiml.comment.domain.CommentRepository;
 import kpl.fiml.comment.dto.*;
 import kpl.fiml.notice.domain.Notice;
 import kpl.fiml.notice.domain.NoticeRepository;
+import kpl.fiml.user.domain.User;
+import kpl.fiml.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,15 +18,17 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CommentService {
 
+    private final UserRepository userRepository;
     private final NoticeRepository noticeRepository;
     private final CommentRepository commentRepository;
 
     @Transactional
-    public CommentCreateResponse create(Long noticeId, CommentCreateRequest request) {
+    public CommentCreateResponse create(Long userId, Long noticeId, CommentCreateRequest request) {
+        User user = getUserById(userId);
         Notice notice = getNoticeById(noticeId);
-        Comment createdComment = commentRepository.save(request.toEntity(notice));
+        Comment createdComment = commentRepository.save(request.toEntity(user, notice));
 
-        return CommentCreateResponse.of(createdComment.getId(), createdComment.getContent(), noticeId);
+        return CommentCreateResponse.of(createdComment.getId(), createdComment.getContent(), userId, noticeId);
     }
 
     public List<CommentDto> findAllByNoticeId(Long noticeId) {
@@ -32,30 +36,45 @@ public class CommentService {
 
         return findList.stream()
                 .filter(comment -> !comment.isDeleted())
-                .map(comment -> CommentDto.of(comment.getId(), comment.getContent(), noticeId, comment.getCreatedAt(), comment.getUpdatedAt()))
+                .map(comment -> CommentDto.of(comment.getId(), comment.getContent(), comment.getUser().getId(), noticeId, comment.getCreatedAt(), comment.getUpdatedAt()))
                 .toList();
     }
 
     @Transactional
-    public CommentUpdateResponse update(Long id, CommentUpdateRequest request) {
+    public CommentUpdateResponse update(Long id, Long userId, CommentUpdateRequest request) {
         Comment findComment = getById(id);
-        findComment.updateContent(request.getContent());
 
-        return CommentUpdateResponse.of(findComment.getId(), findComment.getContent(), findComment.getCreatedAt(), findComment.getUpdatedAt());
+        if (userId.equals(findComment.getUser().getId())) {
+            findComment.updateContent(request.getContent());
+
+            return CommentUpdateResponse.of(findComment.getId(), request.getUserId(), findComment.getContent(), findComment.getCreatedAt(), findComment.getUpdatedAt());
+        } else {
+            throw new IllegalArgumentException("작성자만 수정 가능합니다.");
+        }
     }
 
     @Transactional
-    public CommentDeleteResponse deleteById(Long id) {
+    public CommentDeleteResponse deleteById(Long id, Long userId) {
         Comment findComment = getById(id);
-        findComment.delete();
 
-        return CommentDeleteResponse.of(findComment.getId());
+        if (userId.equals(findComment.getUser().getId())) {
+            findComment.delete();
+            return CommentDeleteResponse.of(findComment.getId());
+        } else {
+            throw new IllegalArgumentException("작성자만 삭제 가능합니다.");
+        }
     }
 
     private Comment getById(Long id) {
         return commentRepository.findById(id)
                 .filter(comment -> !comment.isDeleted())
                 .orElseThrow(() -> new IllegalArgumentException("해당 댓글이 존재하지 않습니다."));
+    }
+
+    private User getUserById(Long userId) {
+        return userRepository.findById(userId)
+                .filter(user -> !user.isDeleted())
+                .orElseThrow(() -> new IllegalArgumentException("해당 사용자가 존재하지 않습니다."));
     }
 
     private Notice getNoticeById(Long noticeId) {
